Added all stages transitions counting.

+++ tests/testthat/test_eeg.R
@@ -6,13 +6,17 @@ test_that("Spectrogram", {
 
   sig <- chirp(seq(-2, 15, by = 0.001), 400, 10, 100, 'quadratic')
 
-  spectrogram(signal = sig,
-                      sRate = 200,
-                      n=2048,
-                      window = 2048,
-                      plot=TRUE)
+  spectrogram(
+    signal = sig,
+    sRate = 200,
+    n=2048,
+    window = 2048,
+    plot=TRUE)
+  
   if(file.exists("Rplots.pdf")){
-    file.remove("Rplots.pdf")
+    if(file.access("Rplots.pdf") == 0){
+      file.remove("Rplots.pdf")
+    }
   }
 
   spec <- spectrogram(signal = sig,
